Add basic api, with pgvector testcontainer
Most (unauthenticated) user interaction is planned to be on hash/...
urls which are precomputed and don't need to load the embedding model.
They are also more cacheable, although this code doesn't provide any
hints for that.

The /import/... is only used in the test right now, I'd much rather
import through the CLI.